add dialog before back to main page

+++ lib/app/modules/lectuter/manageLectuter/views/manage_lectuter_view.dart
@@ -482,6 +482,12 @@ class ManageDataDetail extends StatelessWidget {
               // ),
               ElevatedButton.icon(
                 onPressed: () async {
+                  Get.dialog(
+                    const Center(
+                      child: CircularProgressIndicator(),
+                    ),
+                    barrierDismissible: false,
+                  );
                   controller.selectedIndexFromTable = -1;
                   controller.addressController.selectedProvince.value = '';
                   controller.filePath.value = '';
@@ -493,6 +499,7 @@ class ManageDataDetail extends StatelessWidget {
                   await controller.infoCardController.getSummaryInfo();
                   await controller.lectuterController.listLectuterAffiliate();
                   await controller.lectuterController.listLectuter();
+                  Get.back();
                   Get.toNamed(Routes.LECTUTER);
                 },
                 style: ElevatedButton.styleFrom(
@@ -566,41 +573,43 @@ class ManageDataDetail extends StatelessWidget {
         IconButton(
           icon: const Icon(Icons.delete_sharp),
           onPressed: () async {
-            final result = await Get.dialog(
-              AlertDialog(
-                content: const CustomText(
-                  text: "ยืนยันการลบข้อมูล ?",
-                  scale: 1.2,
-                ),
-                actions: [
-                  TextButton(
-                    child: const CustomText(
-                      text: "ยืนยัน",
-                      scale: 1.2,
-                      color: Colors.red,
-                    ),
-                    onPressed: () => Get.back(result: 'Y'),
+            if (controller.selectedId > 0) {
+              final result = await Get.dialog(
+                AlertDialog(
+                  content: const CustomText(
+                    text: "ยืนยันการลบข้อมูล ?",
+                    scale: 1.2,
                   ),
-                  TextButton(
-                    child: const CustomText(
-                      text: "ปิด",
-                      scale: 1.2,
-                      color: Colors.green,
+                  actions: [
+                    TextButton(
+                      child: const CustomText(
+                        text: "ยืนยัน",
+                        scale: 1.2,
+                        color: Colors.red,
+                      ),
+                      onPressed: () => Get.back(result: 'Y'),
                     ),
-                    onPressed: () => Get.back(result: "N"),
-                  ),
-                ],
-              ),
-            );
-            if (result == 'Y') {
-              Get.dialog(
-                const Center(
-                  child: CircularProgressIndicator(),
+                    TextButton(
+                      child: const CustomText(
+                        text: "ปิด",
+                        scale: 1.2,
+                        color: Colors.green,
+                      ),
+                      onPressed: () => Get.back(result: "N"),
+                    ),
+                  ],
                 ),
-                barrierDismissible: false,
               );
-              await controller.delete();
-              Get.back();
+              if (result == 'Y') {
+                Get.dialog(
+                  const Center(
+                    child: CircularProgressIndicator(),
+                  ),
+                  barrierDismissible: false,
+                );
+                await controller.delete();
+                Get.back();
+              }
             }
           },
         ),
